requirements.txt
langchain>=0.0.267
google-generativeai>=0.3.0
langchain_google_genai>=0.0.1
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
flask>=2.0.0
